Added test_base_unique_id to test_humidifier
