+++ tests/components/vesync/test_humidifier.py
@@ -28,6 +28,7 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError, ServiceValidationError
+from homeassistant.helpers import entity_registry as er
 
 from .common import (
     ENTITY_HUMIDIFIER,
@@ -259,3 +260,14 @@ async def test_set_mode(
         )
     await hass.async_block_till_done()
     method_mock.assert_called_once()
+
+
+async def test_base_unique_id(
+    hass: HomeAssistant,
+    humidifier_config_entry: MockConfigEntry,
+    entity_registry: er.EntityRegistry,
+) -> None:
+    """Test that unique_id is based on subDeviceNo."""
+    # vesync-device.json defines subDeviceNo for 200s-humidifier as 4321.
+    entity = entity_registry.async_get(ENTITY_HUMIDIFIER)
+    assert entity.unique_id.endswith("4321")
